% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{pi}
\alias{pi}
\title{Amino Acid/Nucleotide Diversity (Index Pi)}
\usage{
pi(seqs, count, weighted)
}
\arguments{
\item{seqs}{A character vector containing the distinct haplotype sequences from an individual subject.}

\item{count}{A numeric vector of observed read counts for each haplotype.}

\item{weighted}{A logical value indicating if the haplotypes should be weighted by their frequencies.
If \code{FALSE}, all haplotypes have equal weighting.}
}
\value{
Returns the nucleotide or amino acid diversity as a numeric value.
}
\description{
This function calculates the nucleotide diversity (pi), given an individual's vector of distinct
haplotype sequences and vector of observed read counts for each haplotype. The index can be calculated
using equal weighting for all haplotypes present, or using weights proportional to the frequency of
each haplotype in the sample. Haplotypes may be nucleotide or amino acid sequences.
}
\details{
Nucleotide diversity (pi) is a genetic diversity index that measures the average Hamming distance
(number of nucleotide or amino acid substitutions) between any two sequences of equal length in an
individual's sample. It is a form of a more general metric called Rao's quadratic entropy.
Hamming distance is traditionally used to measure pair-wise differences when calculating
nucleotide diversity, but other methods of calculating distance scores can also be
considered, such as the Levenshtein edit distance or the K80 and F81 substitution models
(Levenshtein 1966; Nei & Kumar 2000). Hamming distance requires the two sequences compared to
be of equal length, but this condition is not necessary for metrics such as the
Levenshtein distance, which accounts for insertions and deletions as well as substitutions.

The unweighted nucleotide diversity is more susceptible than the weighted nucleotide diversity to
sample size bias because rare haplotypes carry more weight. In Gregori et al. (2016), the unweighted
nucleotide diversity is referred to as the nucleotide diversity at the entity level.

Nucleotide diversity is computed as equation 10.5 in Nei (1987) and Equation
IVc in Gregori et al. (2016): \deqn{\pi = n/(n-1) * \sum \sum p_i d_{ij} p_j,}
where \eqn{n} is the sample size, \eqn{p_i} and \eqn{p_i} are the frequencies of two haplotypes, and
\eqn{d_{ij}} is the Hamming distance between the two haplotypes.
}
\examples{
# example with amino acid sequences
seqs <- c("ACDEFGHI","KLMNPQRS","TVWYACDE","SMWLWCAN")
count <- c(10,200,366,14)
weighted <- TRUE
pi(seqs, count, weighted)

# example with nucleotide sequences
seqs <- c("TACCTGGCG","TACTAAGGG", "TACGATGAC")
count <- c(200,50,15)
weighted <- FALSE
pi(seqs, count, weighted)

}
\references{
Gregori, J., Perales, C., Rodriguez-Frias, F., Esteban, J. I., Quer, J., & Domingo, E. (2016).
Viral quasispecies complexity measures. \emph{Virology}, 493, 227-237.

Levenshtein, V. I. (1966, February). Binary codes capable of correcting deletions, insertions, and
reversals. In \emph{Soviet physics doklady} (Vol. 10, No. 8, pp. 707-710).

Nei, M. (1987). \emph{Molecular evolutionary genetics.} Columbia university press.

Nei, M., & Kumar, S. (2000). \emph{Molecular evolution and phylogenetics.} Oxford university press.
}
\seealso{
See \code{\link[Biostrings]{stringDist}} in \strong{Biostrings} for the function that calculates the distance scores.
}
