% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qD.profile.R
\name{qD.profile}
\alias{qD.profile}
\title{Hill Number Profile}
\usage{
qD.profile(count, q = NULL)
}
\arguments{
\item{count}{A numeric vector of observed read counts for each haplotype.}

\item{q}{A numeric vector of the different exponents for which Hill numbers will be calculated.
The default is \code{NULL}, which gives rise to a grid of q values spanning from 0 to 10 and
including \code{Inf}.}
}
\description{
This function plots the Hill number profile for an individual, given a vector of the observed
read counts for each haplotype and a grid of exponents, q. Haplotypes may be
nucleotide or amino acid sequences.
}
\details{
The Hill number profile plots \eqn{qD} vs. \eqn{q} and provides a visual representation
of the diversity within a sample. The height of the curve is a non-increasing function of \eqn{q}
because as \eqn{q} increases, abundant haplotypes are weighted more heavily, so the number of
haplotypes that meet this criteria of abundance decreases. The slope of the curve reflects the
uneveness of the relative abundances of the haplotypes. The more uneven the distribution of
relative abundances, the steeper the curve. In other words, a flat curve corresponds to a
completely even distribution among haplotypes, while a steep curve indicates a few highly
dominant haplotypes and many rare haplotypes.
}
\examples{
count <- c(10,200,366,14)
qD.profile(count)

qD.profile(count, q=NULL)

count <- c(80,200,366,14)
q <- seq(0,3,by=0.1)
qD.profile(count, q)

}
\seealso{
See \code{\link{qD}} for calculating Hill numbers. See \code{\link[vegan]{renyi}} in \strong{vegan}
for plotting Renyi and Hill number accumulation curves.
}
